tests/lua: update SCByteVar tests for suricata.bytevar lib
