+++ tests/lua-byte-extract-pre8/README.md
@@ -1,0 +1,1 @@
+Tests Lua's accessing variables from byte_extract and byte_math in lua match scripts.
