+++ tests/lua-byte-extract/lua-byte-math.lua
@@ -1,11 +1,16 @@
-function init(args)
-    local needs = {}
-    needs["bytevar"] = {"var2"}
-    return needs
+local bytevars = require("suricata.bytevar")
+
+function init(sig)
+    bytevars.map(sig, "var2")
+    return {}
+end
+
+function thread_init()
+    bv2 = bytevars.get("var2")
 end
 
 function match(args)
-    local var2 = SCByteVarGet(0)
+    local var2 = bv2:value()
 
     if var2 and var2 == 0x48545450 then
         return 1
